backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
pydantic
python-multipart
faiss-cpu
sentence-transformers
Pillow
numpy
pytest
